Change menu to match master
