--- apps/admin/components/common/routes.ts
@@ -1,6 +1,0 @@
-export const routes = {
-  dashboard: "/dashboard",
-  users: "/models/users",
-  courses: "/models/courses",
-  accountSettings: "/settings/account",
-};
